// src/app/model/sign-up-model.ts
export default interface SignUpModel{
    id: string;
    name: string;
    userName: string;
    userMobile: number;
    userDOB: Date;
    userAddress: string;
    userState: string;
    userCountry: string;
    userCitizenship: string;
    userCitizenStatus: string;
    userGender: string;
    userDocProof: string;
    userDocNo: string;
    userAccountType: string;
    userBranchNamne: string;
    userDepositAmount: string;
    userRegDate: Date;
    userRefAccHolderName: string;
    userAccHolderAddress: string;
    userGuardianType: string;
    userGuardianName: string;
    userMaritalStatus: string;
}